requirements.txt
pytest>=8.0.0
requests>=2.31.0
allure-pytest>=2.13.5
selene>=2.0.0rc9
curlify>=2.2.1
python-dotenv>=1.0.1